requirements.txt
Django==2.1
pytz==2019.2
djangorestframework=3.11.1